requirements.txt
streamlit==1.31.0
newsapi-python==0.2.7
yfinance==0.2.36
nltk==3.8.1
pandas==2.2.0
matplotlib==3.8.2
numpy==1.26.3
